+++ src/test.zig
@@ -3,4 +3,5 @@ comptime {
     _ = @import("util.zig");
     _ = @import("socket.zig");
     _ = @import("ipc.zig");
+    _ = @import("label.zig");
 }